Reduce code duplication in cfd actors
Extract shared handlers code into a common module
