+++ daemon/src/lib.rs
@@ -1,6 +1,7 @@
 pub mod actors;
 pub mod auth;
 pub mod bitmex_price_feed;
+pub mod cfd_actors;
 pub mod db;
 pub mod fan_out;
 pub mod housekeeping;
